+++ simvue/client.py
@@ -92,8 +92,7 @@ class Client:
             if not value:
                 logger.warning(f"No {label} specified")
 
-        self._headers: dict[str, str] = {
-            "Authorization": f"Bearer {self._user_config.server.token.get_secret_value()}",
+        self._headers: dict[str, str] = self._user_config.headers | {
             "Accept-Encoding": "gzip",
         }