+++ simvue/config/parameters.py
@@ -22,6 +22,9 @@ logger = logging.getLogger(__file__)
 
 
 class ServerSpecifications(pydantic.BaseModel):
+    model_config: typing.ClassVar[pydantic.ConfigDict] = pydantic.ConfigDict(
+        extra="forbid", strict=True
+    )
     url: pydantic.AnyHttpUrl | None
     token: pydantic.SecretStr | None