+++ tests/functional/test_run_execute_process.py
@@ -9,7 +9,7 @@ import filecmp
 
 from simvue import Run, Client
 from simvue.executor import get_current_shell
-from simvue.sender import sender
+from simvue.sender import Sender
 
 @pytest.mark.executor
 def test_monitor_processes(create_plain_run_offline: tuple[Run, dict]):
@@ -24,7 +24,8 @@ def test_monitor_processes(create_plain_run_offline: tuple[Run, dict]):
         _run.add_process(f"process_1_{os.environ.get('PYTEST_XDIST_WORKER', 0)}", Command="Write-Output 'Hello World!'", executable="powershell")
         _run.add_process(f"process_2_{os.environ.get('PYTEST_XDIST_WORKER', 0)}", Command="Get-ChildItem", executable="powershell")
         _run.add_process(f"process_3_{os.environ.get('PYTEST_XDIST_WORKER', 0)}", Command="exit 0", executable="powershell")
-    sender(_run._sv_obj._local_staging_file.parents[1], 1, 10, ["folders", "runs", "alerts"], throw_exceptions=True)
+    _sender = Sender(_run._sv_obj._local_staging_file.parents[1], 1, 10, throw_exceptions=True)
+    _sender.upload(["folders", "runs", "alerts"], )
 
 
 @pytest.mark.executor